feat(boa): in operator
For parsing:

The expression! macro is extended to support a Keyword typed value in its
op arguments. This followed the spec that some expressions are separated
by Punctuators (like == etc.) but others are separated by Keywords (like
this one, or others such as instanceof).

This needed an implementation of PartialEq for Keyword and Punctuator in
order to fix a type error caused by the repeated match guard.

An as_binop method was added to Keyword to mimic the one in Punctuator.

This may be better suited to a trait implementation of TryFrom for the
BinOp enum, with implementations for TryFrom<Keyword> and
TryFrom<Punctuator>.

For interpreting:

Some abstract operations were implemented that are used in the semantics
of the operator in the spec.

An as_object cast method was addedto ValueData in order to get a ref to
that data as an object in order to use the Object methods on a
ValueData value.

Tests:

Add unit tests to interpreter and parser for the new operator.

Solves #88