Update README.md
Tries to make the README.md file more consistent. In some places the lines ended with a ".", while in others it did not. The licenses at the ended were missing some links and some sessions, such as [Benchmark] did not have any text besides the main link, among other things.