Some cleanup
This adds [Prettier][prettier] formatting, and adds checks in the
CI for both, the JS/MD styles and the fact that the wasm target builds.
I also upgraded dependencies for both the Rust crates and the boa WASM
demo.
Finally, I removed the Makefile that we were not using.

I also took the opportunity to add `#[track_caller]` in the parser test
functions, so that we know what test snippet is failing, instead of all
panics pointing to the checker functions. This is a new feature of
Rust 1.46, so this raises the minimum supported Rust version.

[prettier]: https://prettier.io/