Clean-up contribution guidelines, dependencies, Test262, MSRV
This commit updates all dependencies in Rust and Yarn, also cleans-up
the `test_ignore.txt` file a bit, where we had legacy comments.

It uses the new flag in Cargo 1.56 to indicate that the minimum supported
Rust version is, in fact, 1.56, which should help some compatibility
diagnostics.

It also updates the Test262 sub-module with the latest changes to the
official test-suite.

Finally, it adds a clarification in the contribution guidelines to
make users of the Boa tester know that they can get more complete
output by using the "very, very, verbose" output mode.