Interner support in the parser (#1765)
This builds on top of #1758 to try to bring #1763 to life.

Something that should probably be done here would be to convert `JsString` to a `Sym` internally. Then, further optimizations could be done adding common strings to a custom interner type (those that we know statically).

This is definitely work in progress, but I would like to have feedback on the API, and feel free to contribute.

Co-authored-by: raskad <32105367+raskad@users.noreply.github.com>