Only run benchmarks on PRs when a label is set (#2114)
This changes our ci benchmarks to only run when the label `run-benchmark` is set on the PR.

The motivation is to reduce the time waiting on benchmarks to run while working on PRs. Also this saves some ci minutes which is always good.
When we spot changes that we suspect impact performance, we can add the `run-benchmark` label to the PR and the benchmarks will run.