Documentation Updates (#2463)
<!---
Thank you for contributing to Boa! Please fill out the template below, and remove or add any
information as you feel necessary.
--->

Submitting this as a draft for feedback/second opinions. This draft contains some changes to the documentation.

Quick Overview:
  - Potential `Boa` header for Boa's crates added to `boa_engine`.
  - Changes the wording to a lot of module headers (See `builtins` module and `object/builtins` module).
  - Updating built-in wrapper's code examples to use `?` operator.
  - Adds the doc logo URL to a few crates that didn't have it.

The main idea of this draft is to move away from the "This module implements" wording as it feels a bit duplicative when listed under the Modules section (mainly focusing around changes in `boa_engine` to start).

While working on this, I had a question about whether we should be using JavaScript or ECMAScript in the Boa's documentation. We do seem to currently use both, and this draft uses JavaScript heavily in the wording.