Use opcode table rather than match
After decoding an opcode, `match` is used to find a proper `execute`
function for the opcode. But, the `match` may not be able to be
optimized into a linear table jump by rust compiler, so it may use
multiple branches to find the function.

This change makes the opcode to use a table explicitly. This change
improves the benchmark score of Richards about 1-2% (22.8 -> 23.2).