Updated dependencies, removes `remove_dir_all`, which is vulnerable (#2685)
After this, we are still waiting for `indexmap` & `dashmap` to provide the new `hashbrown` to reduce duplicate dependencies, and for `criterion` to remove `clap` and release a new version. We're also waiting for a new version of `icu_datagen` that bumps the `zip` dependency to avoid a potential vulnerability. Ideally, they would also bump the `simple_logger` dependency, which is pretty outdated. In any case, `simple_logger` still uses an unmaintained `atty` dependency.

Relevant issues:
 - https://github.com/xacrimon/dashmap/pull/250
 - https://github.com/unicode-org/icu4x/pull/3150
 - https://github.com/bheisler/criterion.rs/pull/599
 - https://github.com/borntyping/rust-simple_logger/issues/74