Add a js_class to implement the Class trait without boilerplate (#3872)
* Add a js_class to implement the Class trait without boilerplate

This also adds a JsInstance that verifies that `this` is of the proper
class, and an `Ignore` that ignore arguments.

The syntax is a bit special because of limitations of macro_rules. For
example, the way fields are defined. It was impossible to keep both
assignments (e.g. `public field = 123;`) and dynamic fields.

This reduces significantly the boilerplate for declaring classes. The
example in class.rs is more than twice as long (if you remove comments)
than the macro is.

* clippies

* Fix usage of macros

* Fix some imports issues

* Fix some imports issues, again

* Add a way to alias a function, e.g. change case

* Address comment and get rid of unwraps