`setTimeout`, `setInterval` and `clearInterval` (and the same `clearTimeout`) implementations (#4130)
* Make the HostHooks shareable

* `setTimeout`, `setInterval` and `clearInterval` (and the same `clearTimeout`) implementations

This adds non-async implementations of the `setTimeout`/`setInterval` API functions. They are by default registered in the context when registering all APIs from `boa_runtime`.

* Add timeout jobs to tokio/smol event_loop examples and fix borrow issue

* Fix compiling issue in smol_event_loop

* Fix documentation generation for job

* Implement an internal time type and Clock trait

This will add the capacity for the entire engine to mock time during
tests. No impact on performance should be noticeable.

Fixed #4144

* Retype JsDuration in/out to unsigned

A negative duration is invalid.

* Rework the PR to merge the clock from #4149

* Rework the PR to merge the clock from #4149

* Fix smol/tokio event loops

* Remove unused crate