Add a #[boa_class] proc macro attribute
This transforms an `impl XYZ` block into a class. It might not be fully
fool proof, and it doesn't implement everything that can be done in the
Class implementation, but it is sufficient to get started.

This should reduce the code and complexity of class declaration. It
supports accessors, constructors and methods, and uses the same logic
for transforming arguments from `JsValue` that is used by
`IntoJsFunctionCopied`. This makes it very convenient.

It also copy the `impl` block verbatim so methods inside the block can
be used in Rust as regular methods.

There is an additional new functionality that isn't supported by
`js_class!`: you can use `&self` or `&mut self` in your method to get
a borrowed struct instance. That makes it very convenient to declare
regular Rust methods. Unfortunately, it also allows for double borrows
if the method calls anything into JavaScript again. I failed to find
a proper way to communicate this, outside of documentation.