Add a #[boa_class] proc macro attribute (#4271)
* Add a #[boa_class] proc macro attribute

This transforms an `impl XYZ` block into a class. It might not be fully
fool proof, and it doesn't implement everything that can be done in the
Class implementation, but it is sufficient to get started.

This should reduce the code and complexity of class declaration. It
supports accessors, constructors and methods, and uses the same logic
for transforming arguments from `JsValue` that is used by
`IntoJsFunctionCopied`. This makes it very convenient.

It also copy the `impl` block verbatim so methods inside the block can
be used in Rust as regular methods.

There is an additional new functionality that isn't supported by
`js_class!`: you can use `&self` or `&mut self` in your method to get
a borrowed struct instance. That makes it very convenient to declare
regular Rust methods. Unfortunately, it also allows for double borrows
if the method calls anything into JavaScript again. I failed to find
a proper way to communicate this, outside of documentation.

* clippies

* clippies

* Add support for static methods if there is no receiver (and no attr method)

* clippies

* Add support for having &mut Context as an argument

* Add a test for rename a method

* Sometimes the path is not just an identifier. Use argument index instead

* Update core/engine/src/value/mod.rs

Co-authored-by: Haled Odat <8566042+HalidOdat@users.noreply.github.com>

* Update core/engine/src/value/mod.rs

Co-authored-by: Haled Odat <8566042+HalidOdat@users.noreply.github.com>

* Bunch of comments addressed

* Adding tests and fixing bugs, yeehaw

* clippies oops

* Move to from_fn_ptr which seems to work and would be better

* Add more descriptive errors

* Use Ident::parse_any for parsing the argument as static is a keyword, and other things

---------

Co-authored-by: Haled Odat <8566042+HalidOdat@users.noreply.github.com>