Simplify SourcePositionGuard creation
Add a Spanned trait for any types/elements of an AST to be able
to generate a Span, and implement that for any types that are
used in the source position backtrace.

Then added a method to simplify the creation of the source
position. Since it's always created from a ByteCompiler it
makes sense to have it as an associated method.