Introduce async `ModuleLoader`s (#4328)
This PR changes `ModuleLoader` so that it is an async trait instead of a
trait that takes a closure that needs to be executed. This makes things
much nicer for the composing module loaders, since they can compose
nicely using async await syntax.

~One small optimization was to use a `RefModuleLoader` for the composing
loaders instead of only `ModuleLoader`. This is because composers like
`FallbackModuleLoader` are much nicer if they can take their receiver as
`Rc<Self>`, but their inner loaders must take `&Self` to avoid having
something like
`FallbackModuleLoader<L,R>(Rc<L>,Rc<R>)`. This is just an optimization
though, so we could fallback to only having a single trait if you think
having unnecessary `Rc`s inside `Rc`s is a worth tradeoff for API
simplicity.~ Removed in favour of the simpler API.

Finally, a caveat is that I hid the `DynModuleLoader` from the API using
`Any`, so getting the loader is now only possible if you know the full
type of the loader:
```rust
context.module_loader() // private to the crate
context.downcast_module_loader::<SimpleModuleLoader>().unwrap() // public
```
This also allows removing `register_module` and `get_module`, which were
a bit clunky because many modules used different things to index their
internal caches, and most of the loaders didn't have a concept of
"registering" and "getting" modules. Though, I'm not opposed to adding
them back if we decide they're important to have.

cc @hansl since you were very interested in having something like this.