Refactor `JsString`'s static string tagging (#4334)
It changes the following:

- Remove `usize` tagging for known-static strings (remove `tag_ptr`
dependency)
- Instead of storing a `StaticJsString` which points to a `&'static
JsStr<'static>`, we store the tagged `JsStr` static reference directly.
- Remove `StaticJsString` and `RefCount` union.
- Store `Cell<usize>` for refcount in `JsString`
- Change `RAW_STRING_CACHE` to store a `HashSet<&'static
JsStr<'static>>` instead of `FxHashMap<JsStr<'static>, usize>` (reduce
size of key-value pair by 16 bytes)

We tag the `&'static JsStr<'static>` pointer by using
`byte_add()/byte_sub()`, this preserves the strict-provenance., when we
check if we tag we use `addr()` discards the provenance for the tag
check, but that's fine since we don't reconstruct a pointer from the
addr. Tested with `miri` with `-Zmiri-strict-provenance`, and it didn’t
report any problems.