Prevent evalutation of code with `--dump-ast` flag (#4337)
Prevent evaluation of code when `--dump-ast` flag is provided.

Issue being that we are running the code after dumping the AST, so we
are effectively running the code snippet twice (this does not happen for
file execution AST dump). Which causes a duplicate lexical binding
error, that causes the `collect_bindings` to fail.