Some general bug fixes for Temporal implementation (#4349)
<!---
Thank you for contributing to Boa! Please fill out the template below,
and remove or add any
information as you feel necessary.
--->

This Pull Request related to the ongoing work for #1804.

It changes the following:

- Removes outdated `to_temporal_calendar_slot` + related changes
- Updates handling of time fields in `PlainTime` and `PlainDateTime`
ctors
- updated `first` to `get_or_undefined(0)` in
`PlainDate::to_plain_datetime`

Exciting news about this PR. Once merged, we will have only 1 test left
for `PlainDate` and 5 or less for `PlainDateTime` and `PlainTime`.