Temporal documentation update, part 1 (#4353)
<!---
Thank you for contributing to Boa! Please fill out the template below,
and remove or add any
information as you feel necessary.
--->

This Pull Request begins the general updates to temporal to bring it
more up to date with the rest of the built-ins.

For background on this, in the past, our temporal conformance was still
fairly low and the MDN documentation was not released. However, the
Temporal MDN documentation was released earlier this year, and our
conformance overall is approaching 100%, so this PR and subsequent
follow ups are meant to bring the general state of the Temporal
implementation in Boa to better place.

It adds documentation to `Temporal`, `Temporal.Now`, and
`Temporal.PlainDate`.